# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payment QR Service API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/driver-info/storage": {
            "get": {
                "tags": [
                    "Driver-Info"
                ],
                "summary": "получение списков параметров всех платежек",
                "operationId": "ef00ca0a79556f01483aa49df86b60ab",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentSettings"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Get all child organizations of the user"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/driver-info/save": {
            "put": {
                "tags": [
                    "Driver-Info"
                ],
                "summary": "Сохранения данных",
                "operationId": "80a5ed1af908498fc90a458502dda118",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "Ok - данные были обновлены, или записаны",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentSettings"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Successfully save driver info"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ошибка обновление/записи данных.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            },
                                            "nullable": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Failed save driver info"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/driver-info/{paymentMethod:id}/show": {
            "get": {
                "tags": [
                    "Driver-Info"
                ],
                "summary": "Получение всех значений у метода оплаты по [user] [payment_method]",
                "operationId": "6d8678357ec8aa6819e55a0d1292bebf",
                "parameters": [
                    {
                        "name": "paymentMethod:id",
                        "in": "path",
                        "description": "ID Платежа",
                        "required": true,
                        "schema": {
                            "type": "int",
                            "format": "id"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {},
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/PaymentSettings"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Return all info driver by type"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Общая ошибка сервера.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Error server"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "500"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Ошибка авторизации.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message_error": {
                                            "type": "string",
                                            "example": "Не авторизован."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaymentSettings": {
                "properties": {
                    "ykassa": {
                        "properties": {
                            "Shop id": {
                                "description": "Идентификатор магазина",
                                "type": "string",
                                "example": "12345"
                            },
                            "Key": {
                                "description": "Секретный ключ",
                                "type": "string",
                                "example": "secret_key_value"
                            }
                        },
                        "type": "object"
                    },
                    "test": {
                        "properties": {
                            "Api Key": {
                                "description": "API ключ для тестирования",
                                "type": "string",
                                "example": "test_api_key"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}